Implement push force vector calculation; Improve ApplyForce()

+++ vectors/utils/util.go
@@ -8,13 +8,13 @@ import "math"
 // }
 
 // RadToDeg converts radians to degrees
-func RadToDeg(rad float64) int {
-	return int(rad * 180 / math.Pi)
+func RadToDeg(rad float64) float64 {
+	return rad * 180 / math.Pi
 }
 
 // DegToRad converts degrees to radians
-func DegToRad(deg int) float64 {
-	return 2 * math.Pi * (float64(deg) / 360)
+func DegToRad(deg float64) float64 {
+	return 2 * math.Pi * (deg / 360)
 }
 
 // Round performs rounding of a float type to the nearest int
